requirements.txt
asgiref==3.5.2
Django==4.0.6
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.2.0
sqlparse==0.4.2
whitenoise==6.2.0
